[riscv32] add support for bitmanip
This reworks the way the architecture selection is done: depending on
whether the feature `rv32_bitmanip` is enabled or not, a different value
of `-march` will be used. By default, this feature is *enabled* so it
must explicitely be disabled to not use bitmanip, by adding
  --features=-rv32_bitmanip
to bazel's command line.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>