[transition] Remove useless configuration on rule attribute
The `platform_target` rule is already created using `rv_rule` which
sets the `cfg` attribute to make a platform transition on the
outgoing edges of the rule. Therefore it is redundant to further
make the `target` incoming edge of the rule also transition the
platform. For some reason, `attr.label` will make `crt.attr.target`
an array when `cfg` is set instead of just a single element. I
could not find why in the bazel documentation.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>